+++ src/library/zip/R/assertions.R
@@ -1,10 +1,9 @@
-
 is_string <- function(x) {
   is.character(x) && length(x) == 1 && !is.na(x)
 }
 
 is_character <- function(x) {
-  is.character(x) && !any(is.na(x))
+  is.character(x) && !anyNA(x)
 }
 
 is_character_or_null <- function(x) {
@@ -16,6 +15,5 @@ is_flag <- function(x) {
 }
 
 is_count <- function(x) {
-  is.numeric(x) && length(x) == 1 && !is.na(x) &&
-    as.integer(x) == x
+  is.numeric(x) && length(x) == 1 && !is.na(x) && as.integer(x) == x
 }