+++ src/library/zip/src/zip.h
@@ -26,7 +26,8 @@ typedef enum zip_error_codes {
   R_ZIP_ESETZIPPERM  = 14,
   R_ZIP_ECREATE      = 15,
   R_ZIP_EOPENX       = 16,
-  R_ZIP_FILESIZE     = 17
+  R_ZIP_FILESIZE     = 17,
+  R_ZIP_ECREATELINK  = 18
 } zip_error_codes_t;
 
 typedef void zip_error_handler_t(const char *reason, const char *file,