+++ src/library/zip/tools/getzipexe.R
@@ -1,5 +1,4 @@
-
-if(getRversion() < "3.3.0") setInternet2()
+if (getRversion() < "3.3.0") setInternet2()
 
 if (!file.exists("../tools/zip.exe")) {
   download.file(